+++ crates/ntx-builder/src/builder/server/mod.rs
@@ -1,4 +1,4 @@
-use std::sync::{Arc, Mutex};
+use std::sync::Arc;
 
 use miden_node_proto::{
     generated::{
@@ -15,6 +15,7 @@ use miden_objects::{
     note::{Note, Nullifier},
     transaction::TransactionId,
 };
+use tokio::sync::Mutex;
 use tonic::{Request, Response, Status};
 use tracing::info;
 
@@ -60,10 +61,7 @@ impl Api for NtxBuilderApi {
         let notes: Vec<Note> = try_convert(req.note)
             .map_err(|err| Status::invalid_argument(format!("invalid note list: {err}")))?;
 
-        let mut state = self
-            .state
-            .lock()
-            .map_err(|e| Status::internal(format!("Failed to lock state: {e}")))?;
+        let mut state = self.state.lock().await;
 
         state.add_unconsumed_notes(notes);
 
@@ -92,10 +90,7 @@ impl Api for NtxBuilderApi {
                 Status::invalid_argument(format!("error when convertinf input nullifiers: {err}"))
             })?;
 
-        let mut state = self
-            .state
-            .lock()
-            .map_err(|e| Status::internal(format!("failed to lock state: {e}")))?;
+        let mut state = self.state.lock().await;
 
         state.discard_by_nullifiers(&nullifiers);
 
@@ -114,10 +109,8 @@ impl Api for NtxBuilderApi {
             "Received transaction status updates"
         );
 
-        let mut state = self
-            .state
-            .lock()
-            .map_err(|e| Status::internal(format!("failed to lock state: {e}")))?;
+        let mut state = self.state.lock().await;
+
         for tx in request.updates {
             let tx_id: Digest = tx
                 .transaction_id
@@ -139,7 +132,7 @@ impl Api for NtxBuilderApi {
                     "Committed notes notes for transaction"
                 );
             } else {
-                let n = state.rollback_inflight(tx_id.into());
+                let n = state.rollback_inflight(tx_id);
 
                 info!(
                     target: COMPONENT,